Add files via upload
The Avatar Spinning HTML and CSS animation is my take on an  eye-catching and interactive CSS experimental animation that I have crafted in order to further advance my knowledge of both technologies in my most recent front-end web development from scratch learning session. The design strives to provide  two slick aesthetics one located in the index.html and the other one in the index1.html file, the former being colored green yellow; while the latter colored teal. The essence of the design is to have an avatar along with the text of the website to do a spinning animation, as for the avatar I will be implementing that soon as the main essence of this practice project is to be  able to document my present progress in learning both HTML and CSS the fundamentals of full-stack web development.

Note: Feel free to download or fork this repository for yourself and improve them at your spare time, but make sure to give credits to me as it is my work.
Usage Instructions:

1. Download the provided HTML and CSS files from the attached GitHub repository.

2. Include the CSS file in the <head> section of your HTML file using the <link> tag.

3. Add the necessary HTML markup and assign the appropriate class to the element you wish to animate. In this case, assign the "avatar" class to the desired element.

Enjoy the spinning animation effect as the element rotates smoothly and gracefully.